Isbn: int -> long; added some test; Test now use array list of errors instead of string of errors

+++ OnlineLibrary-Backend/src/main/java/ca/mcgill/ecse321/onlinelibrary/controller/OnlineLibraryRestController.java
@@ -20,7 +20,7 @@ public class OnlineLibraryRestController {
 	
 	@PostMapping(value = { "/bookInfo/{title}"})
 	public BookInfoDto createBookInfo(@PathVariable("title") String title, @RequestParam int numberOfPage,  
-			@RequestParam String author, @RequestParam Integer isbn) throws IllegalArgumentException {
+			@RequestParam String author, @RequestParam long isbn) throws IllegalArgumentException {
 		BookInfo bookInfo = service.createBookInfo(title, numberOfPage, author, isbn);
 		return convertToDto(bookInfo);
 	}
